python-backend/requirements.txt
crawl4ai==0.6.3
ollama==0.5.1
rapidfuzz==3.13.0
sentence-transformers==4.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
python-dotenv==1.1.0
agno==1.7.0
plotly==6.2.0
pandas==2.3.0
python-slugify==8.0.4
fastapi==0.116.1
uvicorn==0.35.0